changes to scene, and player functions
UI for targets was added, environment was updated with more platforms and targets. The player now respawns after falling off the level. The platform has a shader to signify the player has not hit all targets. Once the player hits all targets this shader changes color to white, to show the player they can now leave the level.